fix: align ClickHouse schemas with 26.03 schema changes
- clinical_report: add qualityControls Array(String)
- target_prioritisation: rename maxClinicalTrialPhase -> maxClinicalStage
- targets: add drugFromSourceId to chemicalProbes Tuple
- go: full rewrite with actual 26.03 schema (adds regulates, negativelyRegulates, positivelyRegulates, isObsolete; removes name)
- enhancer_to_gene: change start/end from UInt32 to Int64

+++ config/clickhouse/schema/enhancer_to_gene.sql
@@ -26,8 +26,8 @@ CREATE TABLE if not exists enhancer_to_gene_log (
         'Y',
         'MT'
     ),
-    `start` UInt32,
-    `end` UInt32,
+    `start` Int64,
+    `end` Int64,
     `geneId` String,
     `biosampleName` LowCardinality (String),
     `biosampleId` LowCardinality (String),
